% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvMGarch.R
\docType{class}
\name{tvMGarch-class}
\alias{tvMGarch-class}
\title{An S4 class for a nonstationary multivariate class model.}
\description{
A specification class to create an object of a nonstationary multivariate class model reserved for real (empirical) applications. It inherits from \code{simMGarch}.
}
\section{Slots}{

\describe{
\item{\code{out_of_sample_prop}}{Proportion of y to keep for out-of-sample forecasting expressed in \%.}

\item{\code{out_of_sample_y}}{The out of sample y matrix reserved for forecasting and backtesting exercises.}

\item{\code{in_sample_y}}{The in-sample y matrix reserved for estimation (calibration) and change-point detection.}
}}

\examples{
simObj <- new("simMGarch")
simObj@d <- 10
simObj@n <- 1000
simObj@changepoints <- c(250,750)
simObj <- pc_cccsim(simObj)
empirObj <- new("tvMGarch") #simulated, but treated as a real dataset for illustration
empirObj@y <- simObj@y
empirObj@out_of_sample_prop <- 0.1
#empirObj=garch.seg(object=empirObj,do.parallel = 4)##Not run
}
\references{
Cho, Haeran, and Karolos Korkas. "High-dimensional GARCH process segmentation with an application to Value-at-Risk." arXiv preprint arXiv:1706.01155 (2018).
}
